[test/plugin] rename the package of testcases (#3587)
* remove profiles and rename the package

* remove profiles
