+++ test/plugin/scenarios/httpclient-4.3.x-scenario/src/main/java/org/apache/skywalking/apm/testcase/httpclient/CaseServlet.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.apache.skywalking.apm.testcase.httpclient;
+package org.apache.skywalking.apm.testcase.httpclient;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -24,18 +24,37 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.apache.http.HttpEntity;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.util.EntityUtils;
 
-public class ServletForContextPropagate extends HttpServlet {
+public class CaseServlet extends HttpServlet {
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.setContentType("application/json");
-        PrintWriter out = resp.getWriter();
-        out.print("{'test':'test'}");
-        out.flush();
+        CloseableHttpClient httpclient = HttpClients.createDefault();
+
+        HttpGet httpGet = new HttpGet("http://localhost:8080" + req.getContextPath() + "/case/context-propagate");
+        CloseableHttpResponse response1 = httpclient.execute(httpGet);
+        try {
+            HttpEntity entity1 = response1.getEntity();
+            EntityUtils.consume(entity1);
+        } finally {
+            response1.close();
+        }
+
+        PrintWriter printWriter = resp.getWriter();
+        printWriter.write("success");
+        printWriter.flush();
+        printWriter.close();
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         doGet(req, resp);
     }
+
 }
