+++ test/plugin/scenarios/httpclient-4.3.x-scenario/src/main/java/org/apache/skywalking/apm/testcase/httpclient/HealthCheckServlet.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package test.apache.skywalking.apm.testcase.httpclient;
+package org.apache.skywalking.apm.testcase.httpclient;
 
 import java.io.IOException;
 import java.io.PrintWriter;
