+++ test/plugin/scenarios/httpclient-4.3.x-scenario/src/main/webapp/WEB-INF/web.xml
@@ -24,12 +24,12 @@
 
     <servlet>
         <servlet-name>caseServlet</servlet-name>
-        <servlet-class>test.apache.skywalking.apm.testcase.httpclient.CaseServlet</servlet-class>
+        <servlet-class>org.apache.skywalking.apm.testcase.httpclient.CaseServlet</servlet-class>
     </servlet>
 
     <servlet>
         <servlet-name>healthCheck</servlet-name>
-        <servlet-class>test.apache.skywalking.apm.testcase.httpclient.HealthCheckServlet</servlet-class>
+        <servlet-class>org.apache.skywalking.apm.testcase.httpclient.HealthCheckServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
@@ -44,7 +44,7 @@
 
     <servlet>
         <servlet-name>servletForContextPropagate</servlet-name>
-        <servlet-class>test.apache.skywalking.apm.testcase.httpclient.ServletForContextPropagate</servlet-class>
+        <servlet-class>org.apache.skywalking.apm.testcase.httpclient.ServletForContextPropagate</servlet-class>
     </servlet>
 
     <servlet-mapping>
